requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.11.1
frozendict==2.4.6
google-auth==2.40.2
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
multitasking==0.0.11
numpy==2.2.6
oauthlib==3.2.2
pandas==2.2.3
peewee==3.18.1
platformdirs==4.3.8
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yfinance==0.2.61
